requirements.txt
rich
pycryptodome
colorama
bcrypt

